Fixed user.name min length bag in users_factory. Moved test_data for image testing to factories directory
